\name{GetALet-package}
\alias{GetALet-package}
\alias{GetALet}
\docType{package}
\title{
\packageTitle{GetALet}
}
\description{
\packageDescription{GetALet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GetALet}
\packageIndices{GetALet}
The only function in this package is the alphabet function. 
}
\author{
\packageAuthor{GetALet}

Maintainer: \packageMaintainer{GetALet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
